Refactor Dependency handling. Remove unused code. Initial version of manipulator for REST client. Simplify extension shutdown
