+++ core/src/main/java/org/commonjava/maven/ext/manip/ManipulationManager.java
@@ -23,7 +23,6 @@ import org.commonjava.maven.ext.manip.impl.Manipulator;
 import org.commonjava.maven.ext.manip.io.PomIO;
 import org.commonjava.maven.ext.manip.model.Project;
 import org.commonjava.maven.ext.manip.resolver.ExtensionInfrastructure;
-import org.commonjava.maven.ext.manip.resolver.GalleyInfrastructure;
 import org.commonjava.maven.ext.manip.util.ManipulatorPriorityComparator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -153,9 +152,10 @@ public class ManipulationManager
             }
         }
 
-        // Ensure shutdown of executor service
-        if (infrastructure.containsKey("galley")) {
-            ((GalleyInfrastructure) infrastructure.get("galley")).finish();
+        // Ensure shutdown of GalleyInfrastructure Executor Service
+        for (ExtensionInfrastructure e : infrastructure.values())
+        {
+            e.finish();
         }
         logger.info( "Maven-Manipulation-Extension: Finished." );
     }
@@ -171,7 +171,6 @@ public class ManipulationManager
         throws ManipulationException
     {
         session.setProjects( projects );
-
         for ( final Manipulator manipulator : orderedManipulators )
         {
             manipulator.scan( projects, session );
