+++ core/src/main/java/org/commonjava/maven/ext/manip/impl/Version.java
@@ -422,10 +422,12 @@ public class Version
 
     public String getOSGiVersionString()
     {
+        System.out.println ("### isValidOSGI " + isValidOSGi() + " and qualifier " + !hasQualifier());
         if ( isValidOSGi() && !hasQualifier() )
         {
             return originalVersion;
         }
+        System.out.println ("### numericVersion " + numericVersion);
 
         StringBuilder osgiVersion = new StringBuilder();
         if ( numericVersion )
