+++ core/src/main/java/org/commonjava/maven/ext/manip/spi/RemoteDependenciesSPI.java
@@ -1,0 +1,31 @@
+/**
+ *  Copyright (C) 2015 Red Hat, Inc (jcasey@redhat.com)
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *          http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package org.commonjava.maven.ext.manip.spi;
+
+import org.commonjava.maven.atlas.ident.ref.ProjectRef;
+import org.commonjava.maven.atlas.ident.ref.ProjectVersionRef;
+import org.commonjava.maven.atlas.ident.ref.VersionlessArtifactRef;
+import org.commonjava.maven.ext.manip.ManipulationException;
+import org.commonjava.maven.ext.manip.ManipulationSession;
+import org.commonjava.maven.ext.manip.state.State;
+
+import java.util.Map;
+
+public interface RemoteDependenciesSPI
+{
+    Map<? extends ProjectRef, String> load( final State state, final ManipulationSession session )
+            throws ManipulationException;
+}
