+++ core/src/main/java/org/commonjava/maven/ext/manip/state/CommonDependencyState.java
@@ -1,0 +1,90 @@
+/**
+ *  Copyright (C) 2015 Red Hat, Inc (jcasey@redhat.com)
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *          http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package org.commonjava.maven.ext.manip.state;
+
+import org.commonjava.maven.ext.manip.impl.DependencyManipulator;
+
+import java.util.Properties;
+
+/**
+ * Captures configuration relating to dependency alignment from the POMs. Used by {@link DependencyManipulator}.
+ */
+public class CommonDependencyState
+{
+    /**
+     * The String that needs to be prepended a system property to make it a dependencyExclusion.
+     * For example to exclude junit alignment for the GAV (org.groupId:artifactId)
+     * <pre>
+     * <code>-DdependencyExclusion.junit:junit@org.groupId:artifactId</code>
+     * </pre>
+     */
+    public static final String DEPENDENCY_EXCLUSION_PREFIX = "dependencyExclusion.";
+
+    /**
+     * Enables strict checking of non-exclusion dependency versions before aligning to the given BOM dependencies.
+     * For example, <code>1.1</code> will match <code>1.1-rebuild-1</code> in strict mode, but <code>1.2</code> will not.
+     */
+    public static final String STRICT_DEPENDENCIES = "strictAlignment";
+
+    /**
+     * When false, strict version-alignment violations will be reported in the warning log-level, but WILL NOT FAIL THE BUILD. When true, the build
+     * will fail if such a violation is detected. Default value is false.
+     */
+    public static final String STRICT_VIOLATION_FAILS = "strictViolationFails";
+
+    private final boolean overrideTransitive;
+
+    private final boolean overrideDependencies;
+
+    private final boolean strict;
+
+    private final boolean failOnStrictViolation;
+
+    public CommonDependencyState( final Properties userProps )
+    {
+        overrideTransitive = Boolean.valueOf( userProps.getProperty( "overrideTransitive", "true" ) );
+        overrideDependencies = Boolean.valueOf( userProps.getProperty( "overrideDependencies", "true" ) );
+        strict = Boolean.valueOf( userProps.getProperty( STRICT_DEPENDENCIES, "false" ) );
+        failOnStrictViolation = Boolean.valueOf( userProps.getProperty( STRICT_VIOLATION_FAILS, "false" ) );
+    }
+
+    /**
+     * @return whether to override unmanaged transitive dependencies in the build. Has the effect of adding (or not) new entries
+     * to dependency management when no matching dependency is found in the pom. Defaults to true.
+     */
+    public boolean getOverrideTransitive()
+    {
+        return overrideTransitive;
+    }
+
+    /**
+     * @return whether to override managed dependencies in the build. Defaults to true.
+     */
+    public boolean getOverrideDependencies()
+    {
+        return overrideDependencies;
+    }
+
+    public boolean getStrict()
+    {
+        return strict;
+    }
+
+    public boolean getFailOnStrictViolation()
+    {
+        return failOnStrictViolation;
+    }
+}
