+++ core/src/test/java/org/commonjava/maven/ext/manip/impl/ProjectVersionManipulatorTest.java
@@ -636,14 +636,14 @@ public class ProjectVersionManipulatorTest
             final Set<MavenProject> changed = new HashSet<MavenProject>();
             for ( final MavenProject project : projects )
             {
-                if ( applyVersioningChanges( new Project (project.getOriginalModel()), state ) )
+                if ( applyVersioningChanges( new Project (project.getOriginalModel()) ) )
                 {
                     final String v = versionsByGAV.get( ProjectVersionRef.parse( gav( project ) ) );
                     logger.info( project.getName() + " (" + gav( project ) + "): VERSION MODIFIED\n    New version: "
                         + v );
 
                     // this is a bigger model, so only do this if the originalModel was modded.
-                    applyVersioningChanges( new Project (project.getModel()), state );
+                    applyVersioningChanges( new Project (project.getModel()) );
                     changed.add( project );
 
                     if ( v != null )
