+++ integration-test/src/it/dependency-version-manip-child-module/child/src/main/java/test/HelloWorldwithJUnit.java
@@ -1,0 +1,32 @@
+/**
+ * Copyright (C) 2012 Red Hat, Inc. (jcasey@redhat.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package test;
+
+import org.junit.Test;
+
+public class HelloWorldwithJUnit
+{
+    public static void main (String [] args)
+    {
+        System.out.println("hello");
+    }
+
+    @Test
+    public void test()
+    {
+        // Just a dummy method to verify that we can compile again JUnit 4
+    }
+}
