+++ integration-test/src/it/dependency-version-manip-child-module/test.properties
@@ -1,0 +1,21 @@
+#
+# Copyright (C) 2012 Red Hat, Inc. (jcasey@redhat.com)
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+
+# Override the version of JUnit
+# The colon needs to be escaped to prevent the invoker plugin from misinterpreting it
+#dependencyManagement=org.commonjava.maven.ext\:depMgmt1\:1.0
+#overrideTransitive=false
