+++ integration-test/src/it/dependency-version-manip-child-module/verify.groovy
@@ -1,0 +1,36 @@
+/**
+ * Copyright (C) 2012 Red Hat, Inc. (jcasey@redhat.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+def pomFile = new File( basedir, 'pom.xml' )
+def pomChildFile = new File( basedir, 'child/pom.xml' )
+System.out.println( "Slurping POM: ${pomFile.getAbsolutePath()} and ${pomChildFile.getAbsolutePath()}" )
+
+def pom = new XmlSlurper().parse( pomFile )
+def pomChild = new XmlSlurper().parse( pomChildFile )
+
+def dependency = pom.dependencyManagement.dependencies.dependency.find { it.artifactId.text() == "commons-lang" }
+assert dependency != null
+assert dependency.version.text() == "2.5"
+
+// Test overrideTransitive=false
+def junitDependency = pom.dependencyManagement.dependencies.dependency.find { it.artifactId.text() == "junit" }
+assert junitDependency.size() == 0
+
+def childDependency = pomChild.dependencyManagement.dependencies.dependency.find { it.artifactId.text() == "junit" }
+assert childDependency != null
+assert childDependency.version.text() == "4.1"
+*/
