+++ io/src/main/java/org/commonjava/maven/ext/manip/resolver/ExtensionInfrastructure.java
@@ -34,4 +34,5 @@ public interface ExtensionInfrastructure
                final ArtifactRepository localRepository, final Settings settings, final List<String> activeProfiles)
         throws ManipulationException;
 
+    void finish ();
 }
